requirements.txt
python-telegram-bot>=20.7,<22
APScheduler>=3.10
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
aiosqlite>=0.19
pydantic>=2.8
pydantic-settings>=2.4
python-dotenv>=1.0
httpx>=0.27
beautifulsoup4>=4.12
playwright>=1.46
geopy>=2.4
humanize>=4.9
dateparser>=1.2
